+++ app/components/charts/CumulativeBeaconDepositsChart.tsx
@@ -1,9 +1,9 @@
 "use client";
 
 import { SupportedNetwork, TimeRange, Timeline } from "@/app/utils/types";
-import { useDeposits } from "@/app/components/hooks/useDeposits";
+import useDeposits from "@/app/components/hooks/useDeposits";
 import LineChart from "@/app/components/charts/base/LineChart";
-import { useDepositsGrouping } from "@/app/components/hooks/useDepositsGrouping";
+import useDepositsGrouping from "@/app/components/hooks/useDepositsGrouping";
 
 export interface BeaconDepositsChartProps {
   network: SupportedNetwork;
@@ -13,9 +13,9 @@ export interface BeaconDepositsChartProps {
 
 export default function CumulativeBeaconDepositsChart({ network, timeRange, timeline }: BeaconDepositsChartProps) {
   const { data: rawDepositsData, isLoading: rawDepositsLoading } = useDeposits(network, timeline);
-  const { data: depositsData, isLoading: depositsLoading } = useDepositsGrouping(rawDepositsData, timeRange)
+  const { data: depositsData } = useDepositsGrouping(rawDepositsData, timeRange)
 
-  if (!depositsData || !rawDepositsData || depositsLoading || rawDepositsLoading) {
+  if (!depositsData || !rawDepositsData || rawDepositsLoading) {
     return (
       <div className="w-full mx-auto loading-pulse">
         <LineChart
