+++ app/components/charts/CumulativeStrategyDepositsChart.tsx
@@ -2,9 +2,9 @@
 
 import { SupportedNetwork, SupportedToken, TimeRange, Timeline } from "@/app/utils/types";
 import { getNetworkStrategyTokens } from "@/app/utils/constants";
-import { useDeposits } from "@/app/components/hooks/useDeposits";
+import useDeposits from "@/app/components/hooks/useDeposits";
 import LineChart from "@/app/components/charts/base/LineChart";
-import { useDepositsGrouping } from "@/app/components/hooks/useDepositsGrouping";
+import useDepositsGrouping from "@/app/components/hooks/useDepositsGrouping";
 
 export interface CumulativeStrategyDepositsChartProps {
   network: SupportedNetwork;
@@ -14,11 +14,11 @@ export interface CumulativeStrategyDepositsChartProps {
 
 export default function CumulativeStrategyDepositsChart({ network, timeRange, timeline }: CumulativeStrategyDepositsChartProps) {
   const { data: rawDepositsData, isLoading: rawDepositsLoading } = useDeposits(network, timeline);
-  const { data: depositsData, isLoading: depositsLoading } = useDepositsGrouping(rawDepositsData, timeRange)
+  const { data: depositsData } = useDepositsGrouping(rawDepositsData, timeRange)
   
   const networkStrategyTokens = getNetworkStrategyTokens(network);
 
-  if (!depositsData || !rawDepositsData || depositsLoading || rawDepositsLoading) {
+  if (!depositsData || !rawDepositsData || rawDepositsLoading) {
     return (
       <div className="w-full mx-auto loading-pulse">
         <LineChart
