+++ app/components/charts/CumulativeStrategyWithdrawalsChart.tsx
@@ -1,10 +1,10 @@
 "use client";
 
 import { SupportedNetwork, SupportedToken, TimeRange, Timeline } from "@/app/utils/types";
-import { useWithdrawals } from "@/app/components/hooks/useWithdrawals";
+import useWithdrawals from "@/app/components/hooks/useWithdrawals";
 import LineChart from "@/app/components/charts/base/LineChart";
 import { getNetworkStrategyTokens } from "@/app/utils/constants";
-import { useWithdrawalsGrouping } from "@/app/components/hooks/useWithdrawalsGrouping";
+import useWithdrawalsGrouping from "@/app/components/hooks/useWithdrawalsGrouping";
 
 export interface CumulativeStrategyWithdrawalsChartProps {
   network: SupportedNetwork;
@@ -14,11 +14,11 @@ export interface CumulativeStrategyWithdrawalsChartProps {
 
 export default function CumulativeStrategyWithdrawalsChart({ network, timeRange, timeline }: CumulativeStrategyWithdrawalsChartProps) {
   const { data: rawWithdrawalsData, isLoading: rawWithdrawalsLoading } = useWithdrawals(network, timeline);
-  const { data: withdrawalsData, isLoading: withdrawalsLoading } = useWithdrawalsGrouping(rawWithdrawalsData, timeRange)
+  const { data: withdrawalsData } = useWithdrawalsGrouping(rawWithdrawalsData, timeRange)
   
   const networkStrategyTokens = getNetworkStrategyTokens(network);
 
-  if (!withdrawalsData || !rawWithdrawalsData || withdrawalsLoading || rawWithdrawalsLoading) {
+  if (!withdrawalsData || !rawWithdrawalsData || rawWithdrawalsLoading) {
     return (
       <div className="w-full mx-auto loading-pulse">
         <LineChart
